Added details view of obituaries
Added new activity which is started when clicking on one obituary
in list view, it provides an insigth in all obituary view and
provides two buttons, one when clicked shows the cemetery on
google maps and the other adds an entry to the calendar. The
layout and design require some refactoring.